# contracts/openapi/v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SLOPwise legacy API",
    "version": "1.0.0.2026-09-13",
    "description": "Canonical contract of the current Next.js /api/v1 API. Monetary integer fields use the application's current minor-unit convention."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin Next.js server"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/v1/expenses/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/IdPath" }
      ],
      "get": {
        "operationId": "getExpenseV1",
        "tags": ["Expenses"],
        "summary": "Get one expense visible to the current active group member",
        "responses": {
          "200": {
            "description": "Expense",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpenseResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "updateExpenseV1",
        "tags": ["Expenses"],
        "summary": "Replace editable expense fields using the create-expense command shape",
        "description": "cashPayments is accepted by validation but any non-empty value is rejected by the service. Existing linked cash settlements are retained.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ExpenseCommand" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated expense",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpenseResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "operationId": "deleteExpenseV1",
        "tags": ["Expenses"],
        "summary": "Delete an expense",
        "responses": {
          "200": { "$ref": "#/components/responses/EmptyObject" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/v1/groups/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/IdPath" }
      ],
      "get": {
        "operationId": "getGroupV1",
        "tags": ["Groups"],
        "summary": "Get a group and active members",
        "description": "Payee details are nulled for members to whom the current user does not owe money.",
        "responses": {
          "200": {
            "description": "Group",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GroupResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "updateGroupV1",
        "tags": ["Groups"],
        "summary": "Update group name and/or description as group admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateGroupRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated group",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GroupResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "operationId": "deleteGroupV1",
        "tags": ["Groups"],
        "summary": "Delete a zero-balance group as group admin",
        "responses": {
          "200": { "$ref": "#/components/responses/EmptyObject" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/v1/groups/{id}/expenses": {
      "parameters": [
        { "$ref": "#/components/parameters/IdPath" }
      ],
      "get": {
        "operationId": "listGroupExpensesV1",
        "tags": ["Expenses"],
        "summary": "List one page of group expenses",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "One-based page number. The server uses a fixed page size of 30.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100000,
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense page",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpensePageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "operationId": "createExpenseV1",
        "tags": ["Expenses"],
        "summary": "Create an expense and optional linked cash settlements",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ExpenseCommand" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created expense",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpenseResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/v1/groups/{id}/members": {
      "parameters": [
        { "$ref": "#/components/parameters/IdPath" }
      ],
      "post": {
        "operationId": "addGroupMemberV1",
        "tags": ["Groups"],
        "summary": "Add or reactivate a group member as group admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddGroupMemberRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created or reactivated membership",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GroupMemberResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "operationId": "removeGroupMemberV1",
        "tags": ["Groups"],
        "summary": "Deactivate a member or let the current non-admin member leave",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": { "$ref": "#/components/schemas/OpaqueId" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/EmptyObject" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/groups/{id}/requisites": {
      "parameters": [
        { "$ref": "#/components/parameters/IdPath" }
      ],
      "patch": {
        "operationId": "updateGroupRequisitesV1",
        "tags": ["Groups"],
        "summary": "Update the current member's group-specific payee details",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequisitesRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated group-specific payee details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RequisitesResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
...
    "/api/v1/settlements": {
      "post": {
        "operationId": "createSettlementV1",
        "tags": ["Settlements"],
        "summary": "Record a settlement from the current user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSettlementRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created settlement",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SettlementResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "getCurrentUserV1",
        "tags": ["Users"],
        "summary": "Get current user profile",
        "responses": {
          "200": {
            "description": "The user value can be null if the authenticated account row disappeared.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "patch": {
        "operationId": "updateCurrentUserV1",
        "tags": ["Users"],
        "summary": "Update current user profile and default payee details",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProfileRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "Auth.js same-origin session cookie. Secure deployments may use the __Secure- prefixed cookie name."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/OpaqueId" }
      },
...
    },
    "responses": {
      "EmptyObject": {
        "description": "Successful mutation with an empty JSON object",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/EmptyObject" }
          }
        }
      },
      "BadRequest": {
        "description": "Malformed query according to route-specific checks",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/StringErrorEnvelope" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing authenticated Auth.js user",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/StringErrorEnvelope" },
            "examples": {
              "unauthorized": { "value": { "error": "Unauthorized" } }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated principal is not allowed to perform the operation",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/LegacyErrorEnvelope" }
          }
        }
      },
      "NotFound": {
        "description": "Resource is absent, invisible, or invite is invalid",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/LegacyErrorEnvelope" }
          }
        }
      },
      "Conflict": {
        "description": "Domain conflict, duplicate registration, or exhausted Serializable retries",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/LegacyErrorEnvelope" }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Zod request validation error or domain validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/LegacyErrorEnvelope" }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Required external exchange rate is unavailable",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ServiceErrorEnvelope" }
          }
        }
      },
      "InternalServerError": {
        "description": "Explicit JSON fallback for an unexpected service or registration error. Uncaught framework failures can still use the framework's non-contract response.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ServiceErrorEnvelope" }
          }
        }
      }
    },
    "schemas": {
      "OpaqueId": {
        "type": "string",
        "minLength": 1,
        "description": "Current persistence IDs are opaque strings (typically CUIDs)."
      },
...
      "CalendarDate": {
        "type": "string",
        "format": "date",
        "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}$",
        "description": "Strict literal business date. Timestamps are rejected.",
        "examples": ["2026-09-13"]
      },
      "Currency": {
        "type": "string",
        "enum": [
          "RUB", "USD", "EUR", "AMD", "GEL", "TRY", "THB", "AED", "GBP", "JPY",
          "CNY", "CHF", "CZK", "PLN", "HUF", "KZT", "UZS", "BYN", "AZN", "INR"
        ]
      },
...
      "StringErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "ServiceErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ServiceError" }
        }
      },
      "LegacyErrorEnvelope": {
        "description": "The v1 routes expose several incompatible error envelopes.",
        "oneOf": [
          { "$ref": "#/components/schemas/StringErrorEnvelope" },
          { "$ref": "#/components/schemas/ValidationErrorEnvelope" },
          { "$ref": "#/components/schemas/ServiceErrorEnvelope" }
        ]
      },
...
      "Expense": {
        "type": "object",
        "required": [
          "id", "groupId", "paidById", "createdById", "title", "amount", "currency", "amountBase",
          "customRate", "category", "splitType", "date", "notes", "createdAt", "updatedAt", "paidBy",
          "createdBy", "splits", "settlements"
        ],
        "properties": {
          "id": { "$ref": "#/components/schemas/OpaqueId" },
          "groupId": { "$ref": "#/components/schemas/OpaqueId" },
          "paidById": { "$ref": "#/components/schemas/OpaqueId" },
          "createdById": { "$ref": "#/components/schemas/OpaqueId" },
          "title": { "type": "string" },
          "amount": { "$ref": "#/components/schemas/MoneyInteger" },
          "currency": { "type": "string" },
          "amountBase": {
            "oneOf": [
              { "$ref": "#/components/schemas/MoneyInteger" },
              { "type": "null" }
            ]
          },
          "customRate": { "type": ["number", "null"] },
          "category": { "type": ["string", "null"] },
          "splitType": { "type": "string", "enum": ["EQUAL", "EXACT", "PERCENTAGE"] },
          "date": { "$ref": "#/components/schemas/Timestamp" },
          "notes": { "type": ["string", "null"] },
          "createdAt": { "$ref": "#/components/schemas/Timestamp" },
          "updatedAt": { "$ref": "#/components/schemas/Timestamp" },
          "paidBy": { "$ref": "#/components/schemas/UserSummary" },
          "createdBy": { "$ref": "#/components/schemas/UserSummary" },
          "splits": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ExpenseSplit" }
          },
          "settlements": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ExpenseCashSettlement" }
          }
        }
      },
...
      "UpdateGroupRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string", "maxLength": 500 }
        }
      },
      "AddGroupMemberRequest": {
        "type": "object",
        "required": ["userId"],
        "properties": {
          "userId": { "$ref": "#/components/schemas/OpaqueId" }
        }
      },
      "ExpenseSplitCommand": {
        "type": "object",
        "required": ["userId"],
        "properties": {
          "userId": { "$ref": "#/components/schemas/OpaqueId" },
          "amount": { "type": "integer" },
          "percentage": { "type": "integer" }
        }
      },
      "CashPaymentCommand": {
        "type": "object",
        "required": ["userId", "amount"],
        "properties": {
          "userId": { "$ref": "#/components/schemas/OpaqueId" },
          "amount": { "type": "integer", "minimum": 1 }
        }
      },
      "ExpenseCommand": {
        "type": "object",
        "required": ["title", "amount", "date", "paidById", "splitType", "splits"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "amount": { "type": "integer", "minimum": 1, "maximum": 2000000000 },
          "currency": { "allOf": [{ "$ref": "#/components/schemas/Currency" }], "default": "RUB" },
          "customRate": { "type": "number", "exclusiveMinimum": 0, "maximum": 1000000 },
          "category": { "type": "string" },
          "date": { "$ref": "#/components/schemas/CalendarDate" },
          "paidById": { "$ref": "#/components/schemas/OpaqueId" },
          "notes": { "type": "string", "maxLength": 1000 },
          "splitType": { "type": "string", "enum": ["EQUAL", "EXACT", "PERCENTAGE"] },
          "splits": {
            "type": "array",
            "minItems": 1,
            "items": { "$ref": "#/components/schemas/ExpenseSplitCommand" }
          },
          "cashPayments": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CashPaymentCommand" }
          }
        }
      },
      "CreateSettlementRequest": {
        "type": "object",
        "required": ["groupId", "toUserId", "amount", "date"],
        "properties": {
          "groupId": { "$ref": "#/components/schemas/OpaqueId" },
          "toUserId": { "$ref": "#/components/schemas/OpaqueId" },
          "amount": { "type": "integer", "minimum": 1, "maximum": 2000000000 },
          "currency": { "type": "string", "minLength": 3, "maxLength": 3, "default": "RUB" },
          "date": { "$ref": "#/components/schemas/CalendarDate" },
          "notes": { "type": "string", "maxLength": 500 }
        }
      },
      "RequisitesRequest": {
        "type": "object",
        "description": "Empty strings are trimmed and stored as null. Non-empty strings are trimmed before persistence.",
        "properties": {
          "payeeName": { "type": ["string", "null"], "maxLength": 200 },
          "bankName": { "type": ["string", "null"], "maxLength": 100 },
          "payeeAccount": { "type": ["string", "null"], "maxLength": 100 }
        }
      },
      "UpdateProfileRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/RequisitesRequest" },
          {
            "type": "object",
            "properties": {
              "name": { "type": "string", "minLength": 1, "maxLength": 100 },
              "avatarUrl": {
                "type": ["string", "null"],
                "format": "uri",
                "pattern": "^[hH][tT][tT][pP][sS]?://"
              }
            }
          }
        ]
      },
...
      "GroupResponse": {
        "type": "object",
        "required": ["group"],
        "properties": {
          "group": { "$ref": "#/components/schemas/Group" }
        }
      },
      "GroupMemberResponse": {
        "type": "object",
        "required": ["member"],
        "properties": {
          "member": { "$ref": "#/components/schemas/GroupMember" }
        }
      },
      "ExpenseResponse": {
        "type": "object",
        "required": ["expense"],
        "properties": {
          "expense": { "$ref": "#/components/schemas/Expense" }
        }
      },
      "ExpensePageResponse": {
        "type": "object",
        "required": ["expenses", "total", "hasNext"],
        "properties": {
          "expenses": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Expense" }
          },
          "total": { "type": "integer", "minimum": 0 },
          "hasNext": { "type": "boolean" }
        }
      },
...
      "SettlementResponse": {
        "type": "object",
        "required": ["settlement"],
        "properties": {
          "settlement": { "$ref": "#/components/schemas/Settlement" }
        }
      },
...
      "RequisitesResponse": {
        "type": "object",
        "required": ["requisites"],
        "properties": {
          "requisites": {
            "type": "object",
            "required": ["payeeName", "bankName", "payeeAccount"],
            "properties": {
              "payeeName": { "type": ["string", "null"] },
              "bankName": { "type": ["string", "null"] },
              "payeeAccount": { "type": ["string", "null"] }
            }
          }
        }
      },
...
      "ProfileResponse": {
        "type": "object",
        "required": ["user"],
        "properties": {
          "user": {
            "oneOf": [
              { "$ref": "#/components/schemas/Profile" },
              { "type": "null" }
            ]
          }
        }
      },
...
    }
  }
}