# src/app/api/v1/groups/[id]/expenses/route.ts
import { auth } from "@/lib/auth"
import { createExpenseSchema } from "@/lib/validations/expense"
import * as expensesService from "@/services/expenses.service"
import { handleServiceError } from "@/lib/api-errors"
import { NextResponse } from "next/server"
import {
  toExpensePageResponse,
  toExpenseResponse,
} from "@/lib/api/v1/response-mappers"


type Params = { params: Promise<{ id: string }> }

export async function GET(req: Request, { params }: Params) {
  const session = await auth()
  if (!session?.user?.id) return NextResponse.json({ error: "Unauthorized" }, { status: 401 })

  const { id: groupId } = await params
  const url = new URL(req.url)
  const page = Number(url.searchParams.get("page") ?? 1)
  // Верхняя граница защищает от гигантского OFFSET (page=9e15 и т.п.)
  if (!Number.isSafeInteger(page) || page < 1 || page > 100_000) {
    return NextResponse.json({ error: "Invalid page" }, { status: 400 })
  }

  try {
    const result = await expensesService.getGroupExpenses(groupId, session.user.id, page)
    return NextResponse.json(toExpensePageResponse(result))
  } catch (e) {
    return handleServiceError(e)
  }
}

export async function POST(req: Request, { params }: Params) {
  const session = await auth()
  if (!session?.user?.id) return NextResponse.json({ error: "Unauthorized" }, { status: 401 })

  const { id: groupId } = await params
  const body = await req.json().catch(() => null)
  const parsed = createExpenseSchema.safeParse(body)
  if (!parsed.success) {
    return NextResponse.json({ error: parsed.error.flatten() }, { status: 422 })
  }

  try {
    const expense = await expensesService.createExpense(groupId, session.user.id, parsed.data)
    return NextResponse.json(toExpenseResponse(expense), { status: 201 })
  } catch (e) {
    return handleServiceError(e)
  }
}