# src/app/api/v1/groups/[id]/route.ts
import { auth } from "@/lib/auth"
import { updateGroupSchema } from "@/lib/validations/group"
import * as groupsService from "@/services/groups.service"
import { handleServiceError } from "@/lib/api-errors"
import { NextResponse } from "next/server"
import {
  toGroupDetailResponse,
  toGroupResponse,
} from "@/lib/api/v1/response-mappers"

type Params = { params: Promise<{ id: string }> }

export async function GET(_req: Request, { params }: Params) {
  const session = await auth()
  if (!session?.user?.id) return NextResponse.json({ error: "Unauthorized" }, { status: 401 })

  const { id } = await params
  const group = await groupsService.getGroup(id, session.user.id)
  if (!group) return NextResponse.json({ error: "Not found" }, { status: 404 })
  return NextResponse.json(toGroupDetailResponse(group))
}

export async function PATCH(req: Request, { params }: Params) {
  const session = await auth()
  if (!session?.user?.id) return NextResponse.json({ error: "Unauthorized" }, { status: 401 })

  const { id } = await params
  const body = await req.json().catch(() => null)
  const parsed = updateGroupSchema.safeParse(body)
  if (!parsed.success) {
    return NextResponse.json({ error: parsed.error.flatten() }, { status: 422 })
  }

  try {
    const group = await groupsService.updateGroup(id, session.user.id, parsed.data)
    return NextResponse.json(toGroupResponse(group))
  } catch (e) {
    return handleServiceError(e)
  }
}

export async function DELETE(_req: Request, { params }: Params) {
  const session = await auth()
  if (!session?.user?.id) return NextResponse.json({ error: "Unauthorized" }, { status: 401 })

  const { id } = await params
  try {
    await groupsService.deleteGroup(id, session.user.id)
    return NextResponse.json({})
  } catch (e) {
    return handleServiceError(e)
  }
}